Renamed Node and Time and added 1 test case
Add the UnityTestTools and added a simple test case.  The test case
location needs to be looked at.  Also, the UnityTestTools was
unnecessary, because the unit test runner already ships with, so this
was removed during this build.  However, adding it meant I needed to
rename a bunch of files, so Time and Node are now MyTime and MyNode.